+++ test/prom_ex/grafana_agent/config_renderer_test.exs
@@ -8,6 +8,7 @@ defmodule PromEx.GrafanaAgent.ConfigRendererTest do
     test "should generate a configuration yaml file with the correct substitutions", %{tmp_dir: tmp_dir} do
       template_args = %{
         agent_port: "12345",
+        grpc_port: "54321",
         log_level: "error",
         wal_dir: "/tmp/test/wal",
         scrape_interval: "5s",
@@ -28,6 +29,23 @@ defmodule PromEx.GrafanaAgent.ConfigRendererTest do
 
       assert File.exists?(expected_file_path)
       assert File.read!(expected_file_path) == File.read!("#{__DIR__}/expected_output_config.yml")
+    end
+
+    @tag :tmp_dir
+    test "should configure template", %{tmp_dir: tmp_dir} do
+      template_file = Path.join(tmp_dir, "my_template.yml.eex")
+      File.write!(template_file, "foo: <%= @foo %>")
+
+      template_args = %{
+        foo: "this is a foo",
+        template_file: template_file
+      }
+
+      ConfigRenderer.generate_config_file(template_args, tmp_dir)
+      expected_file_path = "#{tmp_dir}/agent.yml"
+
+      assert File.exists?(expected_file_path)
+      assert File.read!(expected_file_path) == "foo: this is a foo"
     end
   end
 end