+++ test/prom_ex/grafana_agent/downloader_test.exs
@@ -5,16 +5,17 @@ defmodule PromEx.GrafanaAgent.DownloaderTest do
 
   alias PromEx.GrafanaAgent.Downloader
 
-  describe "download_grafana_agent/3" do
-    @tag :tmp_dir
-    test "should raise if an invalid arch is provided", %{tmp_dir: tmp_dir} do
-      expected_error =
-        ~s(Invalid GrafanaAgent version provided. Supported version are: ["0.23.0", "0.22.0", "0.21.2", "0.20.1"])
+  describe "download/2" do
+    test "should be able to download all of the listed versions" do
+      OctoFetch.Test.test_all_supported_downloads(Downloader)
+    end
 
-      assert_raise RuntimeError, expected_error, fn ->
-        {"invalid_version"}
-        |> Downloader.download_grafana_agent(tmp_dir, tmp_dir)
-      end
+    @tag :tmp_dir
+    test "should return an error if an invalid version is provided", %{tmp_dir: tmp_dir} do
+      assert capture_log(fn ->
+               assert {:error, "invalid_version is not a supported version"} =
+                        Downloader.download(tmp_dir, override_version: "invalid_version")
+             end) =~ "Failed to download release from GitHub. invalid_version is not a supported version"
     end
 
     @tag :tmp_dir
@@ -22,24 +23,58 @@ defmodule PromEx.GrafanaAgent.DownloaderTest do
       {os, arch} = get_system_arch()
 
       assert capture_log(fn ->
-               assert {:ok, _output_path} =
-                        Downloader.latest_version()
-                        |> Downloader.download_grafana_agent(tmp_dir, tmp_dir)
-             end) =~ "Fetching GrafanaAgent zip archive"
+               assert {:ok, [_output_path], []} = Downloader.download(tmp_dir)
+             end) =~ "Downloading grafana/agent from https://github.com/grafana/agent/releases/download"
 
       assert tmp_dir
              |> File.ls!()
-             |> Enum.sort() == ["agent-#{os}-#{arch}", "agent-#{os}-#{arch}.zip"]
+             |> Enum.sort() == ["agent-#{os}-#{arch}"]
 
       assert capture_log(fn ->
-               assert {:ok, _output_path} =
-                        Downloader.latest_version()
-                        |> Downloader.download_grafana_agent(tmp_dir, tmp_dir)
-             end) =~ "GrafanaAgent zip archive already present"
+               assert :skip = Downloader.download(tmp_dir)
+             end) =~ "GrafanaAgent binary already present"
 
       assert tmp_dir
              |> File.ls!()
-             |> Enum.sort() == ["agent-#{os}-#{arch}", "agent-#{os}-#{arch}.zip"]
+             |> Enum.sort() == ["agent-#{os}-#{arch}"]
+    end
+
+    @tag :tmp_dir
+    test "should download multiple agents in parallel", %{tmp_dir: tmp_dir} do
+      {os, arch} = get_system_arch()
+
+      a_dir = "#{tmp_dir}/a"
+      b_dir = "#{tmp_dir}/b"
+      File.mkdir!(a_dir)
+      File.mkdir!(b_dir)
+
+      downloader_a =
+        Task.async(fn ->
+          capture_log(fn ->
+            assert {:ok, [_output_path], []} = Downloader.download(a_dir)
+          end)
+        end)
+
+      downloader_b =
+        Task.async(fn ->
+          capture_log(fn ->
+            assert {:ok, [_output_path], []} = Downloader.download(b_dir)
+          end)
+        end)
+
+      assert Task.await(downloader_a) =~
+               "Downloading grafana/agent from https://github.com/grafana/agent/releases/download"
+
+      assert Task.await(downloader_b) =~
+               "Downloading grafana/agent from https://github.com/grafana/agent/releases/download"
+
+      assert a_dir
+             |> File.ls!()
+             |> Enum.sort() == ["agent-#{os}-#{arch}"]
+
+      assert b_dir
+             |> File.ls!()
+             |> Enum.sort() == ["agent-#{os}-#{arch}"]
     end
   end
 
@@ -51,6 +86,7 @@ defmodule PromEx.GrafanaAgent.DownloaderTest do
       {{:unix, :darwin}, arch, 64} when arch in ~w(arm aarch64) -> {:darwin, :arm64}
       {{:unix, :darwin}, "x86_64", 64} -> {:darwin, :amd64}
       {{:unix, :linux}, "aarch64", 64} -> {:linux, :arm64}
+      {{:unix, :freebsd}, "amd64", 64} -> {:freebsd, :amd64}
       {{:unix, _osname}, arch, 64} when arch in ~w(x86_64 amd64) -> {:linux, :amd64}
     end
   end