+++ test/prom_ex/plugins/oban_test.exs
@@ -20,12 +20,7 @@ defmodule PromEx.Plugins.ObanTest do
 
     Events.execute_all(:oban)
 
-    metrics =
-      WebApp.PromEx
-      |> PromEx.get_metrics()
-      |> Metrics.sort()
-
-    assert metrics == Metrics.read_expected(:oban)
+    Metrics.assert_prom_ex_metics(WebApp.PromEx, :oban)
   end
 
   describe "event_metrics/1" do