+++ test/prom_ex/plugins/plug_router_test.exs
@@ -25,12 +25,7 @@ defmodule PromEx.Plugins.PlugRouterTest do
     start_supervised!(TestApp.PromEx)
     Events.execute_all(:plug_router)
 
-    metrics =
-      TestApp.PromEx
-      |> PromEx.get_metrics()
-      |> Metrics.sort()
-
-    assert metrics == Metrics.read_expected(:plug_router)
+    Metrics.assert_prom_ex_metics(TestApp.PromEx, :plug_router)
   end
 
   describe "event_metrics/1" do